Rename landlord to central throughout
Reasoning: the authors of the tenancy package refer to 'landlord' as central, so to keep things consistent it seemed appropriate to follow their standards
