+++ resources/central/js/pages/Welcome.vue
