+++ resources/views/central/app.blade.php
@@ -4,7 +4,7 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1">
 
-    @vite(['resources/landlord/js/app.ts', "resources/landlord/js/pages/{$page['component']}.vue"])
+    @vite(['resources/central/js/app.ts', "resources/central/js/pages/{$page['component']}.vue"])
     @routes
     @inertiaHead
 </head>
